requirements.txt
msal>=1.31.0
httpx>=0.27.0
python-dotenv>=1.0.1
streamlit>=1.39.0
pandas>=2.2.0
beautifulsoup4>=4.12.0
